create abstract solar term
